+++ metadata.rb
@@ -4,8 +4,9 @@ maintainer_email 'adrian.herrera@nativex.com'
 license          'All rights reserved'
 description      'Installs/Configures ad-nativex'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '2.1.1'
-supports         'windows', ">= 6.2"
+version          '3.0.0'
+supports         'windows', '>= 6.2'
 
-depends		 "windows"
-depends		 "ohai-nativex"
+depends		 'windows'
+depends		 'ohai-nativex'
+depends    'chef-sugar'