+++ recipes/unjoindomain.rb
@@ -6,13 +6,44 @@
 #
 # All rights reserved - Do Not Redistribute
 #
-
-# Unjoin teamfreeze.com domain
 creds = Chef::EncryptedDataBagItem.load("credentials", "ad")
-ad_nativex_domain "#{node['ad-nativex']['name']}" do
-  action :unjoin
-  retries 3
-  retry_delay 60
-  domain_pass creds["ad_password"]
-  domain_user creds["ad_username"]
+domain = node['ad-nativex']['name']
+
+if centos?
+
+  # Remove machine from domain
+  ruby_block "Un-joining the #{domain} domain" do
+    block do
+      domain_info = `adcli info #{domain}`
+      if domain_info.include? domain
+        cmd = "echo -n #{creds['ad_password']} | adcli delete-computer --domain=#{domain} "\
+          "--login-user=#{creds['ad_username'].split('@')[0]}@#{domain.upcase} --stdin-password"
+        unjoin_domain = `#{cmd}`
+        Chef::Log.info(unjoin_domain)
+      else
+        Chef::Log.error('Could not query domain')
+        raise
+      end
+    end
+    action :run
+    only_if { `klist -k | head`.include? domain.upcase }
+  end
+
+  file "/etc/krb5.keytab" do
+    action :delete
+    only_if { `klist -k | head`.include? domain.upcase }
+  end
+
+elsif windows?
+
+  # Unjoin teamfreeze.com domain
+  creds = Chef::EncryptedDataBagItem.load("credentials", "ad")
+  ad_nativex_domain "#{node['ad-nativex']['name']}" do
+    action :unjoin
+    retries 3
+    retry_delay 60
+    domain_pass creds["ad_password"]
+    domain_user creds["ad_username"]
+  end
+
 end